use the data from the config file if exist

+++ install.sh
@@ -57,25 +57,41 @@ elif [ $(is_installed spamassassin) == 1 ]; then
 	echo "SpamAssassin is already installed, installation aborted"; exit
 fi
 
-read -p "Type hostname: " HOSTNAME
-read -s -p "Type admin's email password: " PASSWORD && echo -e  
-read -e -p "Do you want to install your own ssl certificates? [n/Y] " SSL_INSTALL_OWN 
+if [ "$HOSTNAME" != "" ]; then
+	read -p "Type hostname: " HOSTNAME
+fi
+
+if [ "$PASSWORD" != "" ]; then
+	read -s -p "Type admin's email password: " PASSWORD && echo -e  
+fi
+
+if [ "$SSL_INSTALL_OWN" != "" ]; then
+	read -e -p "Do you want to install your own ssl certificates? [n/Y] " SSL_INSTALL_OWN 
+fi
 
 if [ "$SSL_INSTALL_OWN" == "n"  ] || [ "$SSL_INSTALL_OWN" == "N"  ]; then
 	#by default use dovecot's self-signed certificate
 	SSL_CA_BUNDLE_FILE=/etc/dovecot/dovecot.pem
 	SSL_PRIVATE_KEY_FILE=/etc/dovecot/private/dovecot.pem
 else
-	while [ ! -f "$SSL_CA_BUNDLE_FILE" ]; do
-		read -p "[SSL] CA Bundle file path: " SSL_CA_BUNDLE_FILE
-	done 
-
-	while [ ! -f "$SSL_PRIVATE_KEY_FILE" ]; do
-		read -p "[SSL] Private key file path: " SSL_PRIVATE_KEY_FILE
-	done 
+	if [ "$SSL_CA_BUNDLE_FILE" != "" ]; then
+		while [ ! -f "$SSL_CA_BUNDLE_FILE" ]; do
+			read -p "[SSL] CA Bundle file path: " SSL_CA_BUNDLE_FILE
+		done
+	fi
+ 
+	if [ "$SSL_PRIVATE_KEY_FILE" != "" ]; then
+		while [ ! -f "$SSL_PRIVATE_KEY_FILE" ]; do
+			read -p "[SSL] Private key file path: " SSL_PRIVATE_KEY_FILE
+		done 
+	fi
 fi
 
-read -e -p "Is this installation is on Docker? [N/y] " IS_ON_DOCKER
+
+if [ "$IS_ON_DOCKER" != "" ]; then
+	read -e -p "Is this installation is on Docker? [N/y] " IS_ON_DOCKER
+fi
+
 
 if [ "$IS_ON_DOCKER" == "y"  ] || [ "$IS_ON_DOCKER" == "Y"  ]; then
 	IS_ON_DOCKER=true
